Fix import and add app singlton
